Salary group details changes

+++ app/controllers/salary_group_details_controller.rb
@@ -28,7 +28,6 @@ class SalaryGroupDetailsController < ApplicationController
 
   def new
     @salary_group_detail = SalaryGroupDetail.new
-
     respond_to do |format|
       format.html # new.html.haml
       format.json { render json: @salary_group_detail }
@@ -42,54 +41,61 @@ class SalaryGroupDetailsController < ApplicationController
   def create
     @salary_group_detail = SalaryGroupDetail.new(params[:salary_group_detail])
     emp_detls =  EmployeeDetail.find_all_by_salary_group_id(@salary_group_detail.salary_group_id)
-    #emp_detls.each do |emp_det|
-    #  puts emp_det.id
-    #end
     respond_to do |format|
-      if emp_detls.empty?
-        month_year = params[:salary_group_detail]["effective_month"]
-        eff_date = Date.strptime month_year, '%b/%Y'
-          if @salary_group_detail.save
-            @emp_details = @salary_group_detail.salary_group.employee_details
-            if @emp_details
-              @emp_details.each do |emp_det|
-                SalaryAllotment.create!(:employee_id => emp_det.employee_id, :employee_detail_id => emp_det.id, :effective_date => eff_date.beginning_of_month, :salary_head_id => @salary_group_detail.salary_head_id, :salary_group_detail_id => @salary_group_detail.id, :salary_allotment =>0)
+      if Paymonth.find_by_month_name(params[:salary_group_detail]["effective_month"])
+        if emp_detls.empty?
+          month_year = params[:salary_group_detail]["effective_month"]
+          eff_date = Date.strptime month_year, '%b/%Y'
+            if @salary_group_detail.save
+              @emp_details = @salary_group_detail.salary_group.employee_details
+              if @emp_details
+                @emp_details.each do |emp_det|
+                  SalaryAllotment.create!(:employee_id => emp_det.employee_id, :employee_detail_id => emp_det.id, :effective_date => eff_date.beginning_of_month, :salary_head_id => @salary_group_detail.salary_head_id, :salary_group_detail_id => @salary_group_detail.id, :salary_allotment =>0)
+                end
               end
+              format.html { redirect_to salary_group_details_path(:param1 => params[:salary_group_detail]['salary_group_id']), notice: 'Salary group detail was successfully created.' }
+            else
+              format.html { redirect_to new_salary_group_detail_path(:param1 => params[:salary_group_detail]['salary_group_id']), notice: 'Salary Head has already been taken' }
             end
-            format.html { redirect_to salary_group_details_path(:param1 => params[:salary_group_detail]['salary_group_id']), notice: 'Salary group detail was successfully created.' }
-          else
-            format.html { redirect_to new_salary_group_detail_path(:param1 => params[:salary_group_detail]['salary_group_id']), notice: 'Salary Head has already been taken' }
-          end
+        else
+          format.html { redirect_to new_salary_group_detail_path(:param1 => params[:salary_group_detail]['salary_group_id']), notice: 'Salary Group is already assigned' }
+        end
       else
-        format.html { redirect_to new_salary_group_detail_path(:param1 => params[:salary_group_detail]['salary_group_id']), notice: 'Salary Group is already assigned' }
+        flash[:notice] = 'Selected Effective month is not created.'
+        format.html { redirect_to new_salary_group_detail_path(:param1 => params[:salary_group_detail]['salary_group_id'])}
       end
     end
   end
 
   def update
     effective_month_existence = SalaryGroupDetail.chk_effective_month params[:salary_group_detail]["effective_month"],params[:id]
-    if effective_month_existence.empty?
-      emp_detls =  EmployeeDetail.find_all_by_salary_group_id(params[:salary_group_detail]["salary_group_id"])
-      if emp_detls.empty?
-        @salary_group_detail = SalaryGroupDetail.new(params[:salary_group_detail])
-        if @salary_group_detail.save
-          redirect_to salary_group_details_path(:param1 => params[:salary_group_detail]['salary_group_id']), notice: 'Salary group detail was successfully created.'
+    if Paymonth.find_by_month_name(params[:salary_group_detail]["effective_month"])
+      if effective_month_existence.empty?
+        emp_detls =  EmployeeDetail.find_all_by_salary_group_id(params[:salary_group_detail]["salary_group_id"])
+        if emp_detls.empty?
+          @salary_group_detail = SalaryGroupDetail.new(params[:salary_group_detail])
+          if @salary_group_detail.save
+            redirect_to salary_group_details_path(:param1 => params[:salary_group_detail]['salary_group_id']), notice: 'Salary group detail was successfully created.'
+          else
+            redirect_to new_salary_group_detail_path(:param1 => params[:salary_group_detail]['salary_group_id']), notice: 'Salary Head has already been taken'
+          end
         else
-          redirect_to new_salary_group_detail_path(:param1 => params[:salary_group_detail]['salary_group_id']), notice: 'Salary Head has already been taken'
+          redirect_to edit_salary_group_detail_path(:param1 => params[:salary_group_detail]['salary_group_id']), notice: 'Salary Group is already assigned'
         end
       else
-        redirect_to edit_salary_group_detail_path(:param1 => params[:salary_group_detail]['salary_group_id']), notice: 'Salary Group is already assigned'
+        if @salary_group_detail.update_attributes(params[:salary_group_detail])
+          if params[:salary_group_detail]["calc_type"] != "Every Month"
+            paymonth_id = Paymonth.find_by_month_name(params[:salary_group_detail]["effective_month"]).id
+            EveryMonthCompValue.destroy_all("paymonth_id=#{paymonth_id} and salary_group_id = #{params[:salary_group_detail]['salary_group_id']}")
+          end
+          redirect_to salary_group_details_path(:param1 => params[:salary_group_detail]['salary_group_id']), notice: 'Salary group detail was successfully updated.'
+        else
+          redirect_to edit_salary_group_detail_path(:param1 => params[:salary_group_detail]['salary_group_id'])
+        end
       end
     else
-      if @salary_group_detail.update_attributes(params[:salary_group_detail])
-        if params[:salary_group_detail]["calc_type"] != "Every Month"
-          paymonth_id = Paymonth.find_by_month_name(params[:salary_group_detail]["effective_month"]).id
-          EveryMonthCompValue.destroy_all("paymonth_id=#{paymonth_id} and salary_group_id = #{params[:salary_group_detail]['salary_group_id']}")
-        end
-        redirect_to salary_group_details_path(:param1 => params[:salary_group_detail]['salary_group_id']), notice: 'Salary group detail was successfully updated.'
-      else
-        redirect_to edit_salary_group_detail_path(:param1 => params[:salary_group_detail]['salary_group_id'])
-      end
+      flash[:notice] = 'Selected Effective month is not created.'
+      redirect_to edit_salary_group_detail_path(:param1 => params[:salary_group_detail]['salary_group_id'])
     end
   end
